add buffer test to finalize function (comments)
add buffer around those points that have a coordinate but no
country_coord due to the too low resolution of the rworldxtra
dataset.

the snippet is located in the finalize_country_name function because
it checks, whether the name found by the buffer fits to the original
name. Why? Because we buffer with .5 decimal degrees and often this
leads to multiple country names since their polygons intersect with
the buffer; st_join with argument "largest=TRUE" could avoid this
problem, however its bleeding-edginess might cause installation
problems for non-pro users.